requirements.txt
opencv-contrib-python>=4.5.5.64
pillow>=9.1.0
numpy>=1.22.2
scipy>=1.8.0
tk>=0.1.0